+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.0.0-alpha.93",
-  "contentSha256_12": "b492406b117b",
+  "pkgVersion": "1.0.0-alpha.94",
+  "contentSha256_12": "f32ce658d6b3",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.0.0-alpha.93";
-export const contentSha256_12 = "b492406b117b";
+export const pkgVersion = "1.0.0-alpha.94";
+export const contentSha256_12 = "f32ce658d6b3";
 export const file = "nodalia-cards.bundle.js";